+++ volumes/index.ts
@@ -63,14 +63,14 @@ import smbswap from "./adapters/smbswap";
 import sunswap from "./adapters/sunswap";
 import whaleswap from "./adapters/whaleswap";
 import nomiswap from "./adapters/nomiswap";
-import fstswap from "./adapters/fstswap";
+// import fstswap from "./adapters/fstswap";
 import beethovenX from "./adapters/beethoven-x";
 import defiSwap from "./adapters/defi-swap";
 import wanswapDex from "./adapters/wanswap-dex";
 import solarbeam from "./adapters/solarbeam";
 import tombFinance from "./adapters/tomb-finance";
 import dfxFinance from "./adapters/dfx-finance";
-import frax from "./adapters/frax";
+import fraxSwap from "./adapters/frax-swap";
 import izumiFinance from "./adapters/izumi-finance";
 import tinyman from "./adapters/tinyman";
 import junoswap from "./adapters/junoswap";
@@ -147,7 +147,7 @@ export default {
   solarbeam,
   "tomb-finance": tombFinance,
   "dfx-finance": dfxFinance,
-  frax,
+  "frax-swap": fraxSwap,
   "izumi-finance": izumiFinance,
   tinyman,
   junoswap,
